+++ src/types/Configs.ts
@@ -1,0 +1,4 @@
+export interface LidarrConfig {
+  serverUrl: string;
+  apiKey: string;
+}
