+++ src/types/Genre.ts
@@ -1,0 +1,6 @@
+import { Song } from "./Song";
+
+export type GenreListing = {
+  name: string;
+  songs: Song[];
+};
