+++ src/types/Playlist.ts
@@ -1,0 +1,13 @@
+import { CoverSource } from "./Cover";
+import { Song } from "./Song";
+
+export interface PlaylistBase {
+    id: string;
+    cover: CoverSource;
+    title: string;
+    subtext: string;
+}
+
+export interface Playlist extends PlaylistBase {
+    songs: Song[]
+}
