+++ src/types/index.ts
@@ -1,0 +1,8 @@
+export * from './Server';
+export * from './Cover';
+export * from './Album';
+export * from './Artist';
+export * from './Playlist';
+export * from './Song';
+export * from './Genre';
+export * from './Configs';
